requirements.txt
streamlit
crewai
firecrawl-py
python-dotenv
langchain-openai
pydantic
